feat : LoginMember 인터페이스 변경 및 실제사용자, 익명사용자 구분할 수 있도록 구현 추가
