+++ src/main/java/nextstep/subway/auth/domain/LoginMember.java
@@ -1,28 +1,9 @@
 package nextstep.subway.auth.domain;
 
-public class LoginMember {
-    private Long id;
-    private String email;
-    private Integer age;
+public interface LoginMember {
+    Long getId();
 
-    public LoginMember() {
-    }
+    String getEmail();
 
-    public LoginMember(Long id, String email, Integer age) {
-        this.id = id;
-        this.email = email;
-        this.age = age;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
+    Integer getAge();
 }
