+++ src/main/java/nextstep/subway/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -1,6 +1,7 @@
 package nextstep.subway.auth.ui;
 
 import nextstep.subway.auth.application.AuthService;
+import nextstep.subway.auth.domain.AnonymousMember;
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.auth.infrastructure.AuthorizationExtractor;
 import org.springframework.core.MethodParameter;
@@ -26,6 +27,9 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         String credentials = AuthorizationExtractor.extract(webRequest.getNativeRequest(HttpServletRequest.class));
+        if (credentials == null) {
+            return new AnonymousMember();
+        }
         return authService.findMemberByToken(credentials);
     }
 }
